Move replace tests to a new file test_replace.py
