+++ python/cudf/tests/test_dataframe.py
@@ -177,37 +177,6 @@ def test_series_init_none():
     assert got.split() == expect.split()
 
 
-def test_series_replace():
-    a1 = np.array([0, 1, 2, 3, 4])
-
-    # Numerical
-    a2 = np.array([5, 1, 2, 3, 4])
-    sr1 = Series(a1)
-    sr2 = sr1.replace(0, 5)
-    np.testing.assert_equal(sr2.to_array(), a2)
-
-    # Categorical
-    psr3 = pd.Series(["one", "two", "three"], dtype='category')
-    psr4 = psr3.replace("one", "two")
-    sr3 = Series.from_pandas(psr3)
-    sr4 = sr3.replace("one", "two")
-    pd.testing.assert_series_equal(sr4.to_pandas(), psr4)
-
-    # List input
-    a6 = np.array([5, 6, 2, 3, 4])
-    sr6 = sr1.replace([0, 1], [5, 6])
-    np.testing.assert_equal(sr6.to_array(), a6)
-
-    a7 = np.array([5.5, 6.5, 2, 3, 4])
-    sr7 = sr1.replace([0, 1], [5.5, 6.5])
-    np.testing.assert_equal(sr7.to_array(), a7)
-
-    # Series input
-    a8 = np.array([5, 5, 5, 3, 4])
-    sr8 = sr1.replace(sr1[:3], 5)
-    np.testing.assert_equal(sr8.to_array(), a8)
-
-
 def test_dataframe_basic():
     np.random.seed(0)
     df = DataFrame()
@@ -1508,41 +1477,6 @@ def test_is_monotonic(gdf):
     assert not gdf.index.is_monotonic_decreasing
 
 
-def test_dataframe_replace():
-    # numerical
-    pdf1 = pd.DataFrame({'a': [0, 1, 2, 3], 'b': [0, 1, 2, 3]})
-    gdf1 = DataFrame.from_pandas(pdf1)
-    pdf2 = pdf1.replace(0, 4)
-    gdf2 = gdf1.replace(0, 4)
-    pd.testing.assert_frame_equal(gdf2.to_pandas(), pdf2)
-
-    # categorical
-    pdf4 = pd.DataFrame({'a': ['one', 'two', 'three'],
-                         'b': ['one', 'two', 'three']}, dtype='category')
-    gdf4 = DataFrame.from_pandas(pdf4)
-    pdf5 = pdf4.replace('two', 'three')
-    gdf5 = gdf4.replace('two', 'three')
-    pd.testing.assert_frame_equal(gdf5.to_pandas(), pdf5)
-
-    # list input
-    pdf6 = pdf1.replace([0, 1], [4, 5])
-    gdf6 = gdf1.replace([0, 1], [4, 5])
-    pd.testing.assert_frame_equal(gdf6.to_pandas(), pdf6)
-
-    pdf7 = pdf1.replace([0, 1], 4)
-    gdf7 = gdf1.replace([0, 1], 4)
-    pd.testing.assert_frame_equal(gdf7.to_pandas(), pdf7)
-
-    # dict input:
-    pdf8 = pdf1.replace({'a': 0, 'b': 0}, {'a': 4, 'b': 5})
-    gdf8 = gdf1.replace({'a': 0, 'b': 0}, {'a': 4, 'b': 5})
-    pd.testing.assert_frame_equal(gdf8.to_pandas(), pdf8)
-
-    pdf9 = pdf1.replace({'a': 0}, {'a': 4})
-    gdf9 = gdf1.replace({'a': 0}, {'a': 4})
-    pd.testing.assert_frame_equal(gdf9.to_pandas(), pdf9)
-
-
 @pytest.mark.xfail(reason="null is not supported in gpu yet")
 def test_dataframe_boolean_mask_with_None():
     pdf = pd.DataFrame({'a': [0, 1, 2, 3], 'b': [0.1, 0.2, None, 0.3]})
@@ -1986,32 +1920,3 @@ def test_series_list_nanasnull(nan_as_null):
 
     assert pa.Array.equals(expect, got.to_arrow())
 
-
-@pytest.mark.parametrize('dtype', ['int8', 'int16', 'int32', 'int64',
-                                   'float32', 'float64'])
-@pytest.mark.parametrize(
-    'null_value',
-    [None, np.nan])
-def test_fillna_numerical(dtype, null_value):
-    fill_value = np.random.randint(0, 5)
-    data = np.array([0, 1, null_value, 2, null_value], dtype='float64')
-    sr = Series(data).astype(dtype)
-
-    expect = np.array([0, 1, fill_value, 2, fill_value], dtype=dtype)
-    got = sr.fillna(fill_value).to_array()
-
-    np.testing.assert_equal(expect, got)
-
-
-@pytest.mark.parametrize(
-    'null_value',
-    [None, np.nan])
-def test_fillna_categorical(null_value):
-    data = ['a', 'b', 'a', null_value, 'c']
-    psr = pd.Series(data, dtype='category')
-    sr = Series.from_pandas(psr)
-
-    expect = psr.fillna('a')
-    got = sr.fillna('a')
-
-    assert_eq(expect, got)
